create dataset script and merged fetch_pdbs.py into get_pdbs.py and smiles_db.py into get_ligand_smiles.py
